Refactor public tournament snapshot wiring and route binding
